# rlp: optimize byte array handling